requirements.txt
google-generativeai
python-dotenv
customtkinter
